Logic was backwards for choosing how to add ssh keys
Fixes #3

I was negating the wrong way, so we were adding the stringed (defaulting
to all keys) version when keys were passed, and adding the object
version when keys weren't passed. Not only was this logic backwards, but
it resulted in `undefined` being dumped in the YAML (which fails with
`safeDump`).

This negation was flipped, and tests were added to assert the behavior